# docs/swagger.yaml
definitions:
...
  authentication.RegisterRequest:
    properties:
      customer_id:
        example: 1
        type: integer
        x-order: "0"
      description:
        example: Customer description
        type: string
        x-order: "4"
      name:
        example: John Doe
        type: string
        x-order: "1"
      password:
        example: password123
        type: string
        x-order: "3"
      username:
        example: johndoe
        type: string
        x-order: "2"
    required:
    - customer_id
    - name
    - password
    - username
    type: object
...
  responses.SuccessResponseSwagger-authentication_LoginResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/authentication.LoginResponse'
        description: ข้อมูลที่ต้องการส่งกลับ
        x-order: "4"
      message:
        description: ข้อความสถานะของ response
        example: OK
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 200
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: true
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
...
info:
  contact: {}
  description: For eZVeiw Lite Back Office Project.
  title: eZView Lite Back Office API
  version: "1.0"
paths:
  /v1/auth/register/user/customer:
    post:
      consumes:
      - application/json
      description: |-
        API นี้ใช้สำหรับการลงทะเบียนผู้ใช้ใหม่สำหรับลูกค้าและคืนค่า JWT token ซึ่ง Response จะมีข้อมูลดังนี้:

        - **token**: JWT token ที่ได้รับจากการลงทะเบียน
        - **expires_at**: เวลาที่ token หมดอายุ
        - **expires_after**: ระยะเวลาที่ token หมดอายุหลังจากนี้ (60 นาที)

        Customer ID ได้จาก API /v1/customers/individual หรือ /v1/customers/company

        **หมายเหตุ:** กรุณาใช้ข้อมูลของ Customer ID, User Name และ Password ที่ถูกต้องเพื่อลงทะเบียนผู้ใช้ใหม่
      parameters:
      - description: Register Request
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/authentication.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: การลงทะเบียนสำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-authentication_LoginResponse'
        "400":
          description: การร้องขอไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      summary: ลงทะเบียนผู้ใช้ใหม่สำหรับลูกค้า
      tags:
      - Authentication
  /v1/battery/general:
    get:
      consumes:
      - application/json
      parameters:
      - description: Battery ID
        in: query
        name: battery_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-battery_EZWGetBatteryGeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: Get Battery general information
      tags:
      - Battery
  /v1/battery/insert:
    post:
      consumes:
      - application/json
      description: เพิ่ม Battery ใหม่ลงในระบบ
      parameters:
      - description: ข้อมูล Battery ที่ต้องการเพิ่ม
        in: body
        name: gsmAntenna
        required: true
        schema:
          $ref: '#/definitions/battery.EZWInsertBatteryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-battery_EZWInsertBatteryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: เพิ่มข้อมูล Battery
      tags:
      - Battery
  /v1/battery/search:
    post:
      consumes:
      - application/json
      description: รับ JSON และส่งพารามิเตอร์ไปยังฟังก์ชัน ezw_search_battery
      parameters:
      - description: Payload สำหรับค้นหา Battery
        in: body
        name: searchRequest
        required: true
        schema:
          $ref: '#/definitions/battery.EZWSearchBatteryModelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ผลลัพธ์การค้นหา Battery
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-battery_EZWSearchBatteryModelResponse'
        "400":
          description: ข้อมูลที่ส่งมาไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "401":
          description: ข้อมูลเข้าสู่ระบบไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - Bearer: []
      summary: ค้นหา Battery
      tags:
      - Battery
  /v1/battery/status:
    get:
      consumes:
      - application/json
      description: API นี้ใช้สำหรับดึงข้อมูลสถานะแบตเตอรี่จากฐานข้อมูล
      produces:
      - application/json
      responses:
        "200":
          description: ดึงข้อมูลสำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-battery_EZWBatteryStatusModelResponse'
        "400":
          description: การร้องขอไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "401":
          description: ข้อมูลเข้าสู่ระบบไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: ดึงข้อมูล Battery Status
      tags:
      - Battery
  /v1/battery/update:
    put:
      consumes:
      - application/json
      description: แก้ไขข้อมูล Battery ที่มีอยู่ในระบบ
      parameters:
      - description: ข้อมูล Battery ที่ต้องการอัปเดต
        in: body
        name: gsmAntenna
        required: true
        schema:
          $ref: '#/definitions/battery.EZWUpdateBatteryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-battery_EZWUpdateBatteryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: แก้ไขข้อมูล Battery
      tags:
      - Battery
  /v1/card-reader/brand:
    get:
      consumes:
      - application/json
      description: API นี้ใช้สำหรับดึงข้อมูลแบรนด์ของ Card Reader จากฐานข้อมูล
      produces:
      - application/json
      responses:
        "200":
          description: ดึงข้อมูลแบรนด์ Card Reader สำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-card_reader_EZWCardReaderBrandResponse'
        "400":
          description: ข้อมูลที่ส่งมาไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "401":
          description: ข้อมูลเข้าสู่ระบบไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: ดึงข้อมูล Card Reader Brand
      tags:
      - Card Reader
  /v1/card-reader/general:
    get:
      consumes:
      - application/json
      description: call ezw_get_card_reader_general_by_id($1)
      parameters:
      - description: Card Reader ID
        in: query
        name: card_reader_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-card_reader_EZWCardReaderGeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: Get Card Reader by ID
      tags:
      - Card Reader
  /v1/card-reader/insert:
    post:
      consumes:
      - application/json
      description: เพิ่ม Card Reader ใหม่ลงในระบบ
      parameters:
      - description: ข้อมูล Card Reader ที่ต้องการเพิ่ม
        in: body
        name: cardReader
        required: true
        schema:
          $ref: '#/definitions/card_reader.EZWInsertCardReaderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-card_reader_EZWInsertCardReaderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: เพิ่มข้อมูล Card Reader
      tags:
      - Card Reader
  /v1/card-reader/model:
    get:
      consumes:
      - application/json
      description: API นี้ใช้สำหรับดึงข้อมูลรุ่นของ Card Reader โดยใช้ parent_id (ไอดีของแบรนด์)
      parameters:
      - description: ไอดีของแบรนด์ (parent_id)
        in: query
        name: parent_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ดึงข้อมูลรุ่น Card Reader สำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-card_reader_EZWCardReaderModelResponse'
        "400":
          description: ข้อมูลที่ส่งมาไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "401":
          description: ข้อมูลเข้าสู่ระบบไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: ดึงข้อมูล Card Reader Model
      tags:
      - Card Reader
  /v1/card-reader/search:
    post:
      consumes:
      - application/json
      description: รับ JSON และส่งพารามิเตอร์ไปยังฟังก์ชัน ezw_search_card_readers
      parameters:
      - description: Payload สำหรับค้นหา Card Reader
        in: body
        name: searchRequest
        required: true
        schema:
          $ref: '#/definitions/card_reader.EZWSearchCardReaderModelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ผลลัพธ์การค้นหา Card Reader
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-card_reader_EZWSearchCardReaderModelResponse'
        "400":
          description: ข้อมูลที่ส่งมาไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "401":
          description: ข้อมูลเข้าสู่ระบบไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: ค้นหา Card Reader
      tags:
      - Card Reader
  /v1/card-reader/status:
    get:
      consumes:
      - application/json
      description: API นี้ใช้สำหรับดึงข้อมูลสถานะแบตเตอรี่จากฐานข้อมูล
      produces:
      - application/json
      responses:
        "200":
          description: ดึงข้อมูลสำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-card_reader_EZWCardReaderStatusResponse'
        "400":
          description: การร้องขอไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "401":
          description: ข้อมูลเข้าสู่ระบบไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: ดึงข้อมูล Card Reader Status
      tags:
      - Card Reader
  /v1/card-reader/update:
    put:
      consumes:
      - application/json
      description: แก้ไขข้อมูล Card Reader ที่มีอยู่ในระบบ
      parameters:
      - description: ข้อมูล Card Reader ที่ต้องการอัปเดต
        in: body
        name: cardReader
        required: true
        schema:
          $ref: '#/definitions/card_reader.EZWUpdateCardReaderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-card_reader_EZWUpdateCardReaderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: แก้ไขข้อมูล Card Reader
      tags:
      - Card Reader
  /v1/create-gps:
    post:
      consumes:
      - application/json
      description: |-
        API นี้ใช้สำหรับสร้างข้อมูล GPS ตามเงื่อนไขที่กำหนด

        - **gps_imei**: หมายเลข IMEI ของ GPS
        - **serial_no**: หมายเลข Serial ของ GPS
        - **brand_id**: รหัสยี่ห้อ GPS
        - **model_id**: รหัสรุ่น GPS
        - **status_id**: รหัสสถานะ GPS
        - **remark**: รายละเอียด
      parameters:
      - description: Create GPS Request
        in: body
        name: createRequest
        required: true
        schema:
          $ref: '#/definitions/gps.CreateGPSRequest'
      produces:
      - application/json
      responses:
        "200":
          description: การร้องขอสร้างข้อมูล GPS สำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-gps_CreateGPSResponse'
        "400":
          description: คำขอไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: สร้างข้อมูล GPS ตามเงื่อนไขที่กำหนด
      tags:
      - GPS
  /v1/customers/individual:
    post:
      consumes:
      - multipart/form-data
      description: API นี้ใช้สำหรับสร้างข้อมูลลูกค้าบุคคลทั่วไป
      parameters:
      - description: First Name
        in: formData
        name: first_name
        required: true
        type: string
      - description: Last Name
        in: formData
        name: last_name
        required: true
        type: string
      - description: ID Card
        in: formData
        name: id_card
        required: true
        type: string
      - description: Phone Number
        in: formData
        name: phone_number
        required: true
        type: string
      - description: Address
        in: formData
        name: address
        required: true
        type: string
      - description: Sub District ID
        in: formData
        name: sub_district_id
        required: true
        type: string
      - description: District ID
        in: formData
        name: district_id
        required: true
        type: string
      - description: Province ID
        in: formData
        name: province_id
        required: true
        type: string
      - description: Zip Code
        in: formData
        name: zip_code
        required: true
        type: string
      - description: Images
        in: formData
        name: images
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: สร้างข้อมูลลูกค้าบุคคลทั่วไปสำเร็จ
          schema:
            $ref: '#/definitions/responses.CreateDataResponseSwagger-customers_CreateIndividualCustomerResponse'
        "400":
          description: การร้องขอไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      summary: สร้างข้อมูลลูกค้าบุคคลทั่วไป
      tags:
      - Customers
  /v1/customers/juristic:
    post:
      consumes:
      - multipart/form-data
      description: API นี้ใช้สำหรับสร้างข้อมูลลูกค้านิติบุคคล
      parameters:
      - description: Juristic Name
        in: formData
        name: juristic_name
        required: true
        type: string
      - description: Contact Name
        in: formData
        name: contact_name
        required: true
        type: string
      - description: Contact Lastname
        in: formData
        name: contact_lastname
        required: true
        type: string
      - description: Contact Position
        in: formData
        name: contact_position
        required: true
        type: string
      - description: Tax ID
        in: formData
        name: tax_id
        required: true
        type: string
      - description: Phone Number
        in: formData
        name: phone_number
        required: true
        type: string
      - description: Address
        in: formData
        name: address
        required: true
        type: string
      - description: Sub District ID
        in: formData
        name: sub_district_id
        required: true
        type: string
      - description: District ID
        in: formData
        name: district_id
        required: true
        type: string
      - description: Province ID
        in: formData
        name: province_id
        required: true
        type: string
      - description: Zip Code
        in: formData
        name: zip_code
        required: true
        type: string
      - description: Images
        in: formData
        name: images
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: สร้างข้อมูลลูกค้านิติบุคคลสำเร็จ
          schema:
            $ref: '#/definitions/responses.CreateDataResponseSwagger-customers_CreateJuristicCustomerResponse'
        "400":
          description: การร้องขอไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      summary: สร้างข้อมูลลูกค้านิติบุคคล
      tags:
      - Customers
  /v1/districts:
    get:
      consumes:
      - application/json
      description: |-
        API นี้ใช้สำหรับดึงรายชื่อเขตหรืออำเภอตามจังหวัดที่กำหนด

        - **provinceId**: รหัสจังหวัด
      parameters:
      - default: th
        description: เลือกภาษา (th หรือ en)
        enum:
        - th
        - en
        in: query
        name: lang
        type: string
      - description: รหัสจังหวัด
        in: query
        name: province_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: การร้องขอข้อมูลเขตหรืออำเภอสำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-array_gps_DistrictResponse'
        "400":
          description: คำขอไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: ดึงข้อมูลเขตหรืออำเภอตามจังหวัดที่เลือก
      tags:
      - Dropdowns
  /v1/gps-antenna/general:
    get:
      consumes:
      - application/json
      parameters:
      - description: GPS‑Antenna ID
        in: query
        name: gps_antenna_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-gps_antenna_GpsAntennaGeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: Get GPS‑Antenna general information
      tags:
      - Gps Antenna
  /v1/gps-antenna/insert:
    post:
      consumes:
      - application/json
      description: เพิ่ม Gps Antenna ใหม่ลงในระบบ
      parameters:
      - description: ข้อมูล Gps Antenna ที่ต้องการเพิ่ม
        in: body
        name: gsmAntenna
        required: true
        schema:
          $ref: '#/definitions/gps_antenna.EZWInsertGpsAntennaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-gps_antenna_EZWInsertGpsAntennaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: เพิ่มข้อมูล Gps Antenna
      tags:
      - Gps Antenna
  /v1/gps-antenna/search:
    post:
      consumes:
      - application/json
      description: รับ JSON และส่งพารามิเตอร์ไปยังฟังก์ชัน ezw_search_gps_antenna
      parameters:
      - description: Payload สำหรับค้นหา Gps Antenna
        in: body
        name: searchRequest
        required: true
        schema:
          $ref: '#/definitions/gps_antenna.EZWSearchGpsAntennaModelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ผลลัพธ์การค้นหา Gps Antenna
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-gps_antenna_EZWSearchGpsAntennaModelResponse'
        "400":
          description: ข้อมูลที่ส่งมาไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "401":
          description: ข้อมูลเข้าสู่ระบบไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - Bearer: []
      summary: ค้นหา Gps Antenna
      tags:
      - Gps Antenna
  /v1/gps-antenna/status:
    get:
      consumes:
      - application/json
      description: API นี้ใช้สำหรับดึงข้อมูลสถานะแบตเตอรี่จากฐานข้อมูล
      produces:
      - application/json
      responses:
        "200":
          description: ดึงข้อมูลสำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-gps_antenna_EZWGpsAntennaStatusModelResponse'
        "400":
          description: การร้องขอไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "401":
          description: ข้อมูลเข้าสู่ระบบไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: ดึงข้อมูล GPS Antenna Status
      tags:
      - Gps Antenna
...
  /v1/tracker-bom/battery:
    get:
      consumes:
      - application/json
      description: เรียกฟังก์ชัน ezw_get_battery()
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-array_trackerbom_BatteryResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: ดึงข้อมูล Battery
      tags:
      - Tracker Bom
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"